+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Dagon 0.35.0 - TBD
+------------------
+- **Core**
+  - Global timer functionality. New methods `Application.setTimer`, `Application.cancelTimer`. When the timer expires, it triggers a `Timer` type event (that can be catched by `onTimerEvent` handler) and yields an application-specific user code. Maximum of 1024 simultaneous timers are supported.
+
 Dagon 0.34.1 - 02 Nov, 2025
 ---------------------------
 - **Core**